+++ playground/src/cli/integ.rs
@@ -65,3 +65,4 @@ fn contain(image: Option<String>) -> Result<ExecutionResult> {
     })
 }
 
+