+++ playground/src/cli/mod.rs
@@ -13,3 +13,4 @@ pub mod release;
 pub mod test;
 pub mod valid;
 
+